move receivedWith classes to package 'types'
