+++ src/commonMain/sqldelight/phoenixdb/fr/acinq/phoenix/db/IncomingPayments.sq
@@ -1,5 +1,5 @@
 import fr.acinq.lightning.bin.db.payments.IncomingOriginTypeVersion;
-import fr.acinq.lightning.bin.db.payments.IncomingReceivedWithTypeVersion;
+import fr.acinq.lightning.bin.db.payments.types.IncomingReceivedWithTypeVersion;
 
 -- incoming payments
 CREATE TABLE incoming_payments (
