+++ src/commonTest/kotlin/fr/acinq/lightning/bin/db/DbMigrationTestsCommon.kt
@@ -1,5 +1,6 @@
 package fr.acinq.lightning.bin.db
 
+import fr.acinq.bitcoin.ByteVector32
 import fr.acinq.bitcoin.Chain
 import fr.acinq.bitcoin.PublicKey
 import fr.acinq.lightning.bin.createAppDbDriver
@@ -29,8 +30,8 @@ class DbMigrationTestsCommon {
         val driver = createAppDbDriver(testdir, Chain.Testnet3, PublicKey.fromHex("03be9f16ffcfe10ec7381506601b1b75c9638d5e5aa8c4e7a546573ee09bc68fa2"))
         val database = createPhoenixDb(driver)
         val paymentsDb = SqlitePaymentsDb(database)
-        val payments = paymentsDb.listIncomingPayments(0, Long.MAX_VALUE, 100, 0, true)
-        payments.forEach { println(it) }
+        val payments = paymentsDb.listIncomingPayments(0, Long.MAX_VALUE, Long.MAX_VALUE, 0, true)
+        payments.forEach { println("${it.first.paymentHash} ${it.first.received?.receivedWith}") }
     }
